Remove double quotes from settings menu commands
The settingsmenu used to generate commands such as:

   duskc run_command toggle "Warp"

This used to work fine, but something appears to have changed in more
recent versions of xmenu that causes it to pass the double quotes on
in verbatim, e.g. when executing the command it will run something like:

   duskc run_command toggle \"Warp\"

This will run but won't actually toggle any functionality. This change
removes the double quotes so that the command becomes:

   duskc run_command toggle Warp

This should in principle work fine for both older and newer versions
of xmenu.